Move more common functions to common_linux.go
