+++ pkg/ebpf/common.go
@@ -102,24 +102,6 @@ func verifyKernelFuncs(path string) ([]string, error) {
 	return missing, nil
 }
 
-func isUbuntu(platform string) bool {
-	return strings.Contains(strings.ToLower(platform), "ubuntu")
-}
-
-func isCentOS(platform string) bool {
-	return strings.Contains(strings.ToLower(platform), "centos")
-}
-
-func isRHEL(platform string) bool {
-	p := strings.ToLower(platform)
-	return strings.Contains(p, "redhat") || strings.Contains(p, "red hat") || strings.Contains(p, "rhel")
-}
-
-// isPre410Kernel compares current kernel version to the minimum kernel version(4.1.0) and see if it's older
-func isPre410Kernel(currentKernelCode uint32) bool {
-	return currentKernelCode < stringToKernelCode("4.1.0")
-}
-
 // snakeToCapInitialCamel converts a snake case to Camel case with capital initial
 func snakeToCapInitialCamel(s string) string {
 	n := ""
